+++ src/pages/ReviewListPage/ReviewListPage.tsx
@@ -417,7 +417,10 @@ export default function ReviewListPage() {
     headerName: t('dateAssigned.label'),
     width: 180,
     renderCell: (e: { row: any }) => {
-      return 'unavailable';
+      // TODO retrieve assigned_on from reviewable response once backend updated & update type + mapping
+      return e.row?.proposal?.assignedOn
+        ? presentDate(e.row.proposal?.assignedOn)
+        : t('unavailable');
     }
   };